IPPADI/requirements.txt
Flask
psycopg2-binary
requests
gdal
numpy
Pillow
